Create a global variable no_dashboard
There are times when I don't want the dashboard to show. I was using the
read_from_stdin - but that seemed to be hacky.

So I've added a global variable that can be used to disable the
dashboard.

Usage:

```bash
  nvim -c "let g:no_dashboard=v:true"
```

This could also be moved into an alias. I'm using it in some scripts
that launch Telescope right from the command line and the dashboard was
getting in the way.